chess-ai/requirements.txt
python-chess==1.999
pynput==1.7.6
pywin32==305
selenium==4.8.2